Junk/requirements.txt
Flask==0.10.1
MarkupSafe==0.23
pycrypto==2.6.1
requests==2.9.1
Werkzeug==0.11.4
wheel==0.24.0
